Part out message handler function into module
Starting separating message processing logic from the client main loop.
